requirements.txt
anthropic>=0.39.0
openpyxl>=3.1.5
pyyaml>=6.0
pandas>=2.0
pdfplumber>=0.10
tabulate>=0.9
